Move the hooks to their own file for clarity of what does what
